@keegancsmith 2020-05-27
> I can't find the place where this was being done before. Is this a new
> check?

This is checking we don't register a name twice. It was done before (see
the previous code which had a panic if descriptions[name] was already set)